Reorganize dependency injection so each function has its own file
